+++ usercenter-dao/src/main/java/cn/org/bjca/footstone/usercenter/dao/model/UpdateStatusLog.java
@@ -5,155 +5,155 @@ import java.util.Date;
 
 public class UpdateStatusLog implements Serializable {
 
-  private Integer id;
+    private Integer id;
 
-  private Integer uid;
+    private Long uid;
 
-  private String userType;
+    private String userType;
 
-  private Integer operatorUid;
+    private Integer operatorUid;
 
-  private String operatorAccount;
+    private String operatorAccount;
 
-  private String oldStatus;
+    private String oldStatus;
 
-  private String newStatus;
+    private String newStatus;
 
-  private String description;
+    private String description;
 
-  private String extField1;
+    private String extField1;
 
-  private String extField2;
+    private String extField2;
 
-  private String extField3;
+    private String extField3;
 
-  private String extField4;
+    private String extField4;
 
-  private Integer version;
+    private Integer version;
 
-  private Date createTime;
+    private Date createTime;
 
-  private static final long serialVersionUID = 1L;
+    private static final long serialVersionUID = 1L;
 
-  private String updateSql;
+    private String updateSql;
 
-  public Integer getId() {
-    return id;
-  }
+    public Integer getId() {
+        return id;
+    }
 
-  public void setId(Integer id) {
-    this.id = id;
-  }
+    public void setId(Integer id) {
+        this.id = id;
+    }
 
-  public Integer getUid() {
-    return uid;
-  }
+    public Long getUid() {
+        return uid;
+    }
 
-  public void setUid(Integer uid) {
-    this.uid = uid;
-  }
+    public void setUid(Long uid) {
+        this.uid = uid;
+    }
 
-  public String getUserType() {
-    return userType;
-  }
+    public String getUserType() {
+        return userType;
+    }
 
-  public void setUserType(String userType) {
-    this.userType = userType == null ? null : userType.trim();
-  }
+    public void setUserType(String userType) {
+        this.userType = userType == null ? null : userType.trim();
+    }
 
-  public Integer getOperatorUid() {
-    return operatorUid;
-  }
+    public Integer getOperatorUid() {
+        return operatorUid;
+    }
 
-  public void setOperatorUid(Integer operatorUid) {
-    this.operatorUid = operatorUid;
-  }
+    public void setOperatorUid(Integer operatorUid) {
+        this.operatorUid = operatorUid;
+    }
 
-  public String getOperatorAccount() {
-    return operatorAccount;
-  }
+    public String getOperatorAccount() {
+        return operatorAccount;
+    }
 
-  public void setOperatorAccount(String operatorAccount) {
-    this.operatorAccount = operatorAccount == null ? null : operatorAccount.trim();
-  }
+    public void setOperatorAccount(String operatorAccount) {
+        this.operatorAccount = operatorAccount == null ? null : operatorAccount.trim();
+    }
 
-  public String getOldStatus() {
-    return oldStatus;
-  }
+    public String getOldStatus() {
+        return oldStatus;
+    }
 
-  public void setOldStatus(String oldStatus) {
-    this.oldStatus = oldStatus == null ? null : oldStatus.trim();
-  }
+    public void setOldStatus(String oldStatus) {
+        this.oldStatus = oldStatus == null ? null : oldStatus.trim();
+    }
 
-  public String getNewStatus() {
-    return newStatus;
-  }
+    public String getNewStatus() {
+        return newStatus;
+    }
 
-  public void setNewStatus(String newStatus) {
-    this.newStatus = newStatus == null ? null : newStatus.trim();
-  }
+    public void setNewStatus(String newStatus) {
+        this.newStatus = newStatus == null ? null : newStatus.trim();
+    }
 
-  public String getDescription() {
-    return description;
-  }
+    public String getDescription() {
+        return description;
+    }
 
-  public void setDescription(String description) {
-    this.description = description == null ? null : description.trim();
-  }
+    public void setDescription(String description) {
+        this.description = description == null ? null : description.trim();
+    }
 
-  public String getExtField1() {
-    return extField1;
-  }
+    public String getExtField1() {
+        return extField1;
+    }
 
-  public void setExtField1(String extField1) {
-    this.extField1 = extField1 == null ? null : extField1.trim();
-  }
+    public void setExtField1(String extField1) {
+        this.extField1 = extField1 == null ? null : extField1.trim();
+    }
 
-  public String getExtField2() {
-    return extField2;
-  }
+    public String getExtField2() {
+        return extField2;
+    }
 
-  public void setExtField2(String extField2) {
-    this.extField2 = extField2 == null ? null : extField2.trim();
-  }
+    public void setExtField2(String extField2) {
+        this.extField2 = extField2 == null ? null : extField2.trim();
+    }
 
-  public String getExtField3() {
-    return extField3;
-  }
+    public String getExtField3() {
+        return extField3;
+    }
 
-  public void setExtField3(String extField3) {
-    this.extField3 = extField3 == null ? null : extField3.trim();
-  }
+    public void setExtField3(String extField3) {
+        this.extField3 = extField3 == null ? null : extField3.trim();
+    }
 
-  public String getExtField4() {
-    return extField4;
-  }
+    public String getExtField4() {
+        return extField4;
+    }
 
-  public void setExtField4(String extField4) {
-    this.extField4 = extField4 == null ? null : extField4.trim();
-  }
+    public void setExtField4(String extField4) {
+        this.extField4 = extField4 == null ? null : extField4.trim();
+    }
 
-  public Integer getVersion() {
-    return version;
-  }
+    public Integer getVersion() {
+        return version;
+    }
 
-  public void setVersion(Integer version) {
-    this.version = version;
-  }
+    public void setVersion(Integer version) {
+        this.version = version;
+    }
 
-  public Date getCreateTime() {
-    return createTime;
-  }
+    public Date getCreateTime() {
+        return createTime;
+    }
 
-  public void setCreateTime(Date createTime) {
-    this.createTime = createTime;
-  }
+    public void setCreateTime(Date createTime) {
+        this.createTime = createTime;
+    }
 
-  public String getUpdateSql() {
-    return this.updateSql;
-  }
+    public String getUpdateSql() {
+        return this.updateSql;
+    }
 
-  public void setUpdateSql(String updateSql) {
-    this.updateSql = updateSql;
-  }
+    public void setUpdateSql(String updateSql) {
+        this.updateSql = updateSql;
+    }
 }
